# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Taskosaur API",
    "description": "API for Taskosaur - Task Management Application",
    "version": "1.0.0",
    "contact": {
      "email": "info@taskosaur.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/workspaces": {
      "get": {
        "tags": ["workspaces"],
        "summary": "Get all workspaces",
        "description": "Returns all workspaces the authenticated user has access to",
        "operationId": "getWorkspaces",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of workspaces",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Workspace"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["workspaces"],
        "summary": "Create workspace",
        "description": "Creates a new workspace",
        "operationId": "createWorkspace",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "description": "Workspace to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workspace created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Slug already in use" 
          }
        }
      }
    },
    "/workspaces/{slug}": {
      "get": {
        "tags": ["workspaces"],
        "summary": "Get workspace by slug",
        "description": "Returns a single workspace identified by its slug",
        "operationId": "getWorkspaceBySlug",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Slug of the workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Workspace not found"
          }
        }
      },
      "put": {
        "tags": ["workspaces"],
        "summary": "Update workspace",
        "description": "Updates an existing workspace",
        "operationId": "updateWorkspace",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Slug of the workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Workspace details to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Workspace not found"
          },
          "409": {
            "description": "Slug already in use"
          }
        }
      },
      "delete": {
        "tags": ["workspaces"],
        "summary": "Delete workspace",
        "description": "Deletes a workspace",
        "operationId": "deleteWorkspace",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Slug of the workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Workspace deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Workspace not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateWorkspaceRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Marketing Team"
          },
          "description": {
            "type": "string",
            "example": "Workspace for the marketing department"
          },
          "slug": {
            "type": "string",
            "example": "marketing-team"
          }
        }
      },
      "UpdateWorkspaceRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Updated Marketing Team"
          },
          "description": {
            "type": "string", 
            "example": "Updated description for the marketing department"
          },
          "avatar": {
            "type": "string",
            "format": "url",
            "example": "https://example.com/avatar.png"
          }
        }
      },
...
      "Workspace": {
        "type": "object",
        "required": ["id", "slug", "name", "description", "projectCount", "memberCount", "recentActivity"],
        "properties": {
          "id": {
            "type": "string",
            "example": "ws-abc"
          },
          "slug": {
            "type": "string",
            "example": "my-cool-workspace"
          },
          "name": {
            "type": "string",
            "example": "My Cool Workspace"
          },
          "description": {
            "type": "string",
            "example": "This is a description of the workspace."
          },
          "projectCount": {
            "type": "integer",
            "example": 5
          },
          "memberCount": {
            "type": "integer",
            "example": 12
          },
          "recentActivity": {
            "type": "string",
            "example": "Task 'Deploy to Prod' updated 2 hours ago"
          },
          "avatar": {
            "type": "string",
            "format": "url",
            "nullable": true,
            "example": "https://example.com/workspace-avatar.png"
          }
        }
      },
...
    }
  }
}